Fetch pull requests and Codex heads exactly once
- `reload`'s own `page = 0` reset re-entered the lookahead
  observer on fresh models, racing two `gh pr list` fetches and
  flashing an empty list; the observer now ignores same-value and
  never-fetched states
- A Codex rollout head cut inside a multi-byte character made the
  whole session vanish and re-read 256KB every poll; the decode
  now drops the partial trailing bytes and unparseable heads
  cache as sentinels
- Each dashboard poll walked the whole Codex session tree once
  per worktree; one listing per root now serves a refresh pass
