+++ Sources/AgentIDEData/CodexTranscriptIndex.swift
@@ -10,6 +10,21 @@ import Synchronization
 struct CodexTranscriptIndex {
     // MARK: Internal
 
+    /// Decodes a fixed-size head read, dropping the final bytes when
+    /// the cut landed inside a multi-byte character; the partial
+    /// trailing line is skipped by the parse anyway.
+    static func decodeHead(_ head: Data) -> String? {
+        var bytes = head
+        for _ in 0 ..< utf8CharacterBytes {
+            if let text = String(bytes: bytes, encoding: .utf8) {
+                return text
+            }
+
+            bytes = bytes.dropLast()
+        }
+        return nil
+    }
+
     /// Sessions under `root` whose embedded working directory
     /// matches, newest first. The session id comes from the file's
     /// metadata line: the rollout file name is not the resume id.
@@ -59,15 +74,29 @@ struct CodexTranscriptIndex {
         let text: String?
     }
 
+    /// One whole-tree listing per root within a refresh pass: the
+    /// dashboard poll asks once per worktree in quick succession
+    /// and the tree cannot meaningfully change between those asks.
+    private struct Listing {
+        let listedAt: ContinuousClock.Instant
+        let entries: [String: Entry]
+    }
+
     /// Heads are enough for the metadata and the first typed user
     /// message, which sits after Codex's large injected instruction
     /// and world-state lines; whole rollout files are larger still.
     private static let headBytes = 262_144
 
+    /// A UTF-8 character spans at most this many bytes.
+    private static let utf8CharacterBytes = 4
+
     /// The most sessions kept indexed, newest first.
     private static let entryCap = 2_000
 
     private static let cache: Mutex<[String: Entry]> = .init([:])
+
+    private static let listings: Mutex<[String: Listing]> = .init([:])
+    private static let listingLifetime: Duration = .seconds(1)
 
     /// Reads the file head: the metadata line names the session and
     /// its working directory, and the first user message titles it.
@@ -78,7 +107,7 @@ struct CodexTranscriptIndex {
 
         defer { try? handle.close() }
         guard let head = try? handle.read(upToCount: headBytes),
-              let text = String(bytes: head, encoding: .utf8)
+              let text = decodeHead(head)
         else {
             return nil
         }
@@ -141,8 +170,14 @@ struct CodexTranscriptIndex {
     }
 
     /// Every session file under the root, parsed or served from the
-    /// cache when unchanged.
+    /// caches when unchanged or recently listed.
     private func indexedEntries(root: String) -> [String: Entry] {
+        let now = ContinuousClock.now
+        if let listing = Self.listings.withLock({ $0[root] }),
+           now - listing.listedAt < Self.listingLifetime {
+            return listing.entries
+        }
+
         let manager = FileManager.default
         var results = [String: Entry]()
         let enumerator = manager.enumerator(atPath: root)
@@ -159,8 +194,12 @@ struct CodexTranscriptIndex {
             let cached = Self.cache.withLock { $0[path] }
             if let cached, cached.modifiedAt == modifiedAt {
                 results[path] = cached
-            } else if let parsed = Self.parseHead(path: path, modifiedAt: modifiedAt) {
-                results[path] = parsed
+            } else {
+                // Unparseable heads cache as an empty sentinel, so
+                // a bad file is not re-read on every poll; the empty
+                // working directory can never match a worktree.
+                results[path] = Self.parseHead(path: path, modifiedAt: modifiedAt)
+                    ?? Entry(modifiedAt: modifiedAt, workingDirectory: "", sessionID: "", title: "")
             }
         }
         // The cache mirrors disk exactly and caps at the newest
@@ -172,6 +211,7 @@ struct CodexTranscriptIndex {
             bounded = Dictionary(uniqueKeysWithValues: Array(newest))
         }
         Self.cache.withLock { $0 = bounded }
+        Self.listings.withLock { $0[root] = Listing(listedAt: now, entries: bounded) }
         return bounded
     }
 }
